+++ amd/src/form_users_selector.js
@@ -16,7 +16,7 @@
 /**
  * Provides the required functionality for an autocomplete element to select a user.
  *
- * @module      local_entities/form_entities_selector
+ * @module      local_shopping_cart/form_users_selector
  * @copyright   2022 Thomas Winkler
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -33,42 +33,42 @@ import {render as renderTemplate} from "core/templates";
  * @param {Function} failure A function to call in case of failure, receiving the error message.
  */
 export async function transport(selector, query, callback, failure) {
-  const request = {
-    methodname: "local_shopping_cart_search_users",
-    args: {
-      query: query,
-    },
-  };
+    const request = {
+        methodname: "local_shopping_cart_search_users",
+        args: {
+            query: query,
+        },
+    };
 
-  try {
-    const response = await Ajax.call([request])[0];
+    try {
+        const response = await Ajax.call([request])[0];
 
-    let labels = [];
+        let labels = [];
 
-    // eslint-disable-next-line no-console
-    console.log(response);
+        // eslint-disable-next-line no-console
+        console.log(response);
 
-    if (response.warnings.length > 0) {
-        callback(response.warnings);
-    } else {
-        response.list.forEach((user) => {
-            labels.push(
-                renderTemplate(
-                "local_shopping_cart/form-user-selector-suggestion",
-                user
-                )
-            );
+        if (response.warnings.length > 0) {
+            callback(response.warnings);
+        } else {
+            response.list.forEach((user) => {
+                labels.push(
+                    renderTemplate(
+                        "local_shopping_cart/form-user-selector-suggestion",
+                        user
+                    )
+                );
             });
             labels = await Promise.all(labels);
 
             response.list.forEach((entity, index) => {
-            entity.label = labels[index];
+                entity.label = labels[index];
             });
             callback(response.list);
+        }
+    } catch (e) {
+        failure(e);
     }
-  } catch (e) {
-    failure(e);
-  }
 }
 
 /**
@@ -79,9 +79,9 @@ export async function transport(selector, query, callback, failure) {
  * @return {Array} New array of the selector options.
  */
 export function processResults(selector, results) {
-  if (!Array.isArray(results)) {
-    return results;
-  } else {
-    return results.map((result) => ({value: result.id, label: result.label}));
-  }
+    if (!Array.isArray(results)) {
+        return results;
+    } else {
+        return results.map((result) => ({value: result.id, label: result.label}));
+    }
 }