+++ db/access.php
@@ -34,6 +34,14 @@ $capabilities = [
                         'manager' => CAP_ALLOW
                 ],
         ],
+        'local/shopping_cart:cashtransfer' => [
+                'riskbitmask' => RISK_PERSONAL,
+                'captype' => 'write',
+                'contextlevel' => CONTEXT_SYSTEM,
+                'archetypes' => [
+                        'manager' => CAP_ALLOW
+                ],
+        ],
         'local/shopping_cart:cashiermanualrebook' => [
                 'riskbitmask' => RISK_PERSONAL,
                 'captype' => 'write',