+++ download_cash_report.php
@@ -67,7 +67,7 @@ $table->define_headers([
     get_string('gateway', 'local_shopping_cart'),
     get_string('orderid', 'local_shopping_cart'),
     get_string('annotation', 'local_shopping_cart'),
-    get_string('usermodified', 'local_shopping_cart')
+    get_string('cashier', 'local_shopping_cart')
 ]);
 
 // Columns.