+++ report.php
@@ -181,7 +181,7 @@ $headers = [
     get_string('gateway', 'local_shopping_cart'),
     get_string('orderid', 'local_shopping_cart'),
     get_string('annotation', 'local_shopping_cart'),
-    get_string('usermodified', 'local_shopping_cart')
+    get_string('cashier', 'local_shopping_cart')
 ];
 // Columns.
 $columns = [